Refactored list display to support more scenarios.
Updated ReceiptDisplayRow and ReceiptList to handle different sets of
columns for different views, and somewhat hackishly implemented new
column selections for the "Snoozed Cases" and "Resolved Cases" views.

Also passed through the correct snoozed-days value from SnoozeForm, so
that we don't have an embarrassing disconnect between the user action
and the displayed result.